Inventory x Controls OTP
I basically tied the controls to the inventory as they were supposed to.
In the debug scene, pick up the three tools, cycle through by holding z
or c, use the tool with e or shift+e if it has multiple uses. Was having
some issues with it on my computer, hopefully will be fixed tomorrow and
we can put in the animations.